flash write bug

+++ driver/eeprom.c
@@ -25,7 +25,7 @@
 #include "ARMCM0.h"
 
 
-void EEPROM_ReadBuffer(uint32_t Address, void *pBuffer, uint8_t Size) {
+void EEPROM_ReadBuffer(uint32_t Address, void *pBuffer, uint16_t Size) {
 
     __disable_irq();
     I2C_Start();
