+++ driver/eeprom.h
@@ -19,7 +19,7 @@
 
 #include <stdint.h>
 
-void EEPROM_ReadBuffer(uint32_t Address, void *pBuffer, uint8_t Size);
+void EEPROM_ReadBuffer(uint32_t Address, void *pBuffer, uint16_t Size);
 
 void EEPROM_WriteBuffer(uint32_t Address, const void *pBuffer, uint8_t WRITE_SIZE);
